Fix the models to sit flush on the XY plane.
The skin() function will twist the layers, so that the first layer is
above/below the XY plane a little from left to right. This means you
can't print them easily like this and can be fixed by producing more of
the outer shell and snipping it off at the XY plane exactly.