Allow for corner radius on length and width
Without this change the length (y dimension) of the stem is `3+dia/2` which is greater than the choc stem recess of `3mm`. See http://www.kailh.com/en/Products/Ks/CS/319.html for reference.

This also tidies the total width to not be reliant on the corner diameter either. (True width was `wids*2 + dia`)